Minor changes.
- Added cudaDeviceSynchronize() before timers for accurate timing.
- Switched back to double precision.
- Removed unnecessary consts.

+++ projects2020/group03/src/cuda/include/common_types.h
@@ -1,3 +1,3 @@
 #pragma once
 
-using realType = float;
+using realType = double;
